コントロラーでの削除の制御

+++ app/controllers/items_controller.rb
@@ -36,7 +36,9 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    item.destroy
+    if user_signed_in? && current_user.id == item.user_id
+      item.destroy
+    end
     redirect_to root_path
   end
